LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/capturing-snap-positions-expected.txt


FAIL The third item should be snapped to by default, not the second's child. assert_equals: expected 200 but got 100
